create exception handler

+++ src/main/java/com/junior/expensemanager/service/ExpenseService.java
@@ -4,6 +4,7 @@ import com.junior.expensemanager.dto.ExpenseDTO;
 import com.junior.expensemanager.dto.ExpenseFilterDTO;
 import com.junior.expensemanager.entity.Expense;
 import com.junior.expensemanager.entity.User;
+import com.junior.expensemanager.exception.ExpenseNotFoundException;
 import com.junior.expensemanager.repository.ExpenseRepository;
 import com.junior.expensemanager.util.DateTimeUtil;
 import org.modelmapper.ModelMapper;
@@ -15,6 +16,7 @@ import java.sql.Date;
 import java.text.DecimalFormat;
 import java.text.ParseException;
 import java.util.List;
+import java.util.Optional;
 import java.util.UUID;
 import java.util.stream.Collectors;
 
@@ -45,16 +47,18 @@ public class ExpenseService {
 //        expenseDTO.setDate(expense.getDate());
 //        return expenseDTO;
 
-         ExpenseDTO expenseDTO = modelMapper.map(expense, ExpenseDTO.class);
-         expenseDTO.setDateString(DateTimeUtil.convertToString(expense.getDate()));
-         return expenseDTO;
+        ExpenseDTO expenseDTO = modelMapper.map(expense, ExpenseDTO.class);
+        expenseDTO.setDateString(DateTimeUtil.convertToString(expense.getDate()));
+        return expenseDTO;
     }
 
     private Expense mapToEntity(ExpenseDTO expenseDTO) throws ParseException {
         // map the DTO to entity
         Expense expense = modelMapper.map(expenseDTO, Expense.class);
         // TODO: generate the expense id
-        expense.setExpenseId(UUID.randomUUID().toString());
+        if (expense.getId() == null) {
+            expense.setExpenseId(UUID.randomUUID().toString());
+        }
         // TODO: set the expense date
         expense.setDate(DateTimeUtil.convertStringToDate(expenseDTO.getDateString()));
         // Return expense entity
@@ -67,13 +71,19 @@ public class ExpenseService {
         expenseRepository.save(expense);
     }
 
-    public void deleteExpenseByExpenseId(String expenseId) throws ParseException {
+    public void deleteExpenseByExpenseId(String expenseId) {
         Expense expense = expenseRepository.findByExpenseId(expenseId);
+        if(expense == null) {
+            throw new ExpenseNotFoundException("Expense not found for the id: "+expenseId);
+        }
         expenseRepository.delete(expense);
     }
 
-    public ExpenseDTO findByIdExpenseId(String expenseId) throws ParseException {
-        Expense expense =  expenseRepository.findByExpenseId(expenseId);
+    public ExpenseDTO findByIdExpenseId(String expenseId) {
+        Expense expense = expenseRepository.findByExpenseId(expenseId);
+        if(expense == null) {
+            throw new ExpenseNotFoundException("Expense not found for the id:"+expenseId);
+        }
         return mapToDTO(expense);
     }
 
@@ -86,14 +96,14 @@ public class ExpenseService {
         Date fromDate = !fromDateStr.isEmpty() ? DateTimeUtil.convertStringToDate(fromDateStr) : new Date(0);
         Date toDate = !toDateStr.isEmpty() ? DateTimeUtil.convertStringToDate(toDateStr) : new Date(System.currentTimeMillis());
         User user = userService.getLoggedInUser();
-        List<Expense> expenses=  expenseRepository.findByNameContainingAndDateBetweenAndUserId(
+        List<Expense> expenses = expenseRepository.findByNameContainingAndDateBetweenAndUserId(
                 keyword,
                 fromDate,
                 toDate,
                 user.getId());
         List<ExpenseDTO> expensesDTO = expenses.stream().map(this::mapToDTO).collect(Collectors.toList());
-        if(sortBy != null) {
-            if(sortBy.equals("date")) {
+        if (sortBy != null) {
+            if (sortBy.equals("date")) {
                 expensesDTO.sort((o1, o2) -> o2.getDate().compareTo(o1.getDate()));
             } else {
                 expensesDTO.sort((o1, o2) -> o2.getAmount().compareTo(o1.getAmount()));
@@ -105,7 +115,7 @@ public class ExpenseService {
     public String totalExpenses(List<ExpenseDTO> expenses) throws ParseException {
         DecimalFormat decimalFormat = new DecimalFormat("#,###.00");
         BigDecimal sum = new BigDecimal(0);
-        for(ExpenseDTO o : expenses) {
+        for (ExpenseDTO o : expenses) {
             sum = sum.add(o.getAmount());
         }
         return decimalFormat.format(sum);
